refactor: rename to bot-framework
