+++ packages/bot-framework/.dockerignore
