+++ packages/bot-framework/.eslintignore
