+++ packages/bot-framework/.eslintrc.json
