+++ packages/bot-framework/Dockerfile
