+++ packages/bot-framework/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "foo",
-  "version": "1.0.0",
+  "name": "@google-automations/bot-framework",
+  "version": "0.0.0",
   "description": "",
   "private": true,
   "author": "Google LLC.",
