+++ packages/bot-framework/src/background/cloud-storage-payload-cache.ts
