+++ packages/bot-framework/src/errors.ts
