+++ packages/bot-framework/src/installations.ts
