+++ packages/bot-framework/src/logging/error-logging.ts
