+++ packages/bot-framework/src/logging/gcf-logger.ts
