+++ packages/bot-framework/src/middleware/logging.ts
